paramétrage total NBDSRV ; à tester (périphérique2 non modifié, conserver 32)

+++ frontal2/src/pageshtml.h
@@ -15,8 +15,8 @@ void remoteHtml(EthernetClient* cli);
 void thermoCfgHtml(EthernetClient* cli);
 void thermoShowHtml(EthernetClient* cli);
 void timersHtml(EthernetClient* cli);
-void detServHtml(EthernetClient* cli,uint32_t* mds,char* lib);
-void detServHtml(EthernetClient* cli,char* buf,char* jsbuf,uint16_t* lb,uint16_t lb0,uint32_t* mds,char* lib);
+void detServHtml(EthernetClient* cli,uint8_t* mds,char* lib);
+void detServHtml(EthernetClient* cli,char* buf,char* jsbuf,uint16_t* lb,uint16_t lb0,uint8_t* mds,char* lib);
 void cfgDetServHtml(EthernetClient* cli);
 int  scalcTh(int bd);
